backend/requirements.txt
django==5.1.1
argon2-cffi==23.1.0
gunicorn==23.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2.binary==2.9.9
